# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CeetFeedback API",
        "description": "API documentation for CeetFeedback",
        "contact": {
            "email": "blezanatanael@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "V1"
    },
    "paths": {
...
        "/api/v1/modules": {
            "get": {
                "tags": [
                    "Modules"
                ],
                "summary": "Récupérer la liste des modules",
                "operationId": "462aac4e505b9bb732211a32a58a8de2",
                "responses": {
                    "200": {
                        "description": "Liste des modules récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ModuleSchema"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Modules"
                ],
                "summary": "Créer un nouveau module",
                "operationId": "30048d853b1b20bc9f2b5fa9f03bd6e0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ModuleSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Module créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Module créé avec succès"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ModuleSchema"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/modules/{id}": {
            "put": {
                "tags": [
                    "Modules"
                ],
                "summary": "Mettre à jour un module",
                "operationId": "7ff19d6036dbf786c36454606cb6f8fe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ModuleSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Module mis à jour avec succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Modules"
                ],
                "summary": "Supprimer un module (soft delete)",
                "operationId": "f6a42ade7b0f06a2f09a2935fda702f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Module supprimé avec succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/modules/{module}/phases": {
            "get": {
                "tags": [
                    "Modules"
                ],
                "summary": "Récupérer toutes les phases d'un module",
                "operationId": "2c1bae68ca0648aa20a127248fc00786",
                "parameters": [
                    {
                        "name": "module",
                        "in": "path",
                        "description": "ID du module",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des phases du module avec count",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "module_id": {
                                            "type": "integer"
                                        },
                                        "count": {
                                            "type": "integer"
                                        },
                                        "phases": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PhaseSchema"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Modules"
                ],
                "summary": "Attacher une phase à un module",
                "operationId": "b198d03e894c01ce18603d73aa8a12dc",
                "parameters": [
                    {
                        "name": "module",
                        "in": "path",
                        "description": "ID du module",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phase_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Phase attachée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Phase attachée au module avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Modules"
                ],
                "summary": "Détacher une phase d'un module",
                "operationId": "7e906ec8fa5e4c386b0367cf5f71157b",
                "parameters": [
                    {
                        "name": "module",
                        "in": "path",
                        "description": "ID du module",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phase_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Phase détachée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Phase détachée du module avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ModuleSchema": {
                "title": "Module",
                "description": "Ressource module",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "slug": {
                        "type": "string",
                        "example": "maths-xyz123"
                    },
                    "name": {
                        "type": "string",
                        "example": "Mathématiques"
                    },
                    "description": {
                        "type": "string",
                        "example": "Module sur les bases des mathématiques."
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-18T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-18T12:34:56Z"
                    }
                },
                "type": "object"
            },
            "PhaseSchema": {
                "title": "Phase",
                "description": "Ressource phase",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "slug": {
                        "type": "string",
                        "example": "phase-abc123"
                    },
                    "name": {
                        "type": "string",
                        "example": "Phase de conception"
                    },
                    "description": {
                        "type": "string",
                        "example": "Cette phase couvre la planification et la conception du projet."
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-18T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-18T12:34:56Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}